+++ lib/utils/chunkWhile.test.ts
@@ -1,21 +1,25 @@
-import { expect, test } from 'vitest';
+import { describe, expect, test } from 'vitest';
 import chunkWhile from './chunkWhile';
 
-test('chunkWhile', () => {
+describe('chunkWhile', () => {
   const source = ['a', 'b', 'a', 1];
 
-  const resultChunk = chunkWhile(source, (item, _index, chunk) => {
-    return chunk.indexOf(item) > -1;
+  test('chunk', () => {
+    const result = chunkWhile(source, (item, _index, chunk) => {
+      return chunk.indexOf(item) > -1;
+    });
+
+    expect(result).toEqual([
+      ['a', 'b'],
+      ['a', 1],
+    ]);
   });
 
-  expect(resultChunk).toEqual([
-    ['a', 'b'],
-    ['a', 1],
-  ]);
+  test('index', () => {
+    const result = chunkWhile(source, (_item, index) => {
+      return index > 1;
+    });
 
-  const resultIndex = chunkWhile(source, (_item, index) => {
-    return index > 1;
+    expect(result).toEqual([['a', 'b'], ['a'], [1]]);
   });
-
-  expect(resultIndex).toEqual([['a', 'b'], ['a'], [1]]);
 });